Added sequence alignment metric that combines the global and local metrics.

+++ src/ehupatras/webrecommendation/sequencealignment/SequenceAlignment.java
@@ -1,5 +1,5 @@
 package ehupatras.webrecommendation.sequencealignment;
 
 public interface SequenceAlignment {
-	public int getScore(String[] seqA, String[] seqB);
+	public float getScore(String[] seqA, String[] seqB);
 }
